requirements.txt
Django==3.2.10
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
Pillow==7.1.1
PyJWT==1.7.1
